afpstats: expose hostname for each afpd child process
store server hostname in the child process data structure
and expose it in the afpstats D-Bus message for printing

introduce an option in afpstats, --hostname, that filters
the output by hostname; also add a --version option for good measure

+++ include/atalk/server_ipc.h
@@ -12,7 +12,7 @@
 #define IPC_LOGINDONE        4
 
 extern int ipc_server_read(server_child_t *children, int fd);
-extern int ipc_child_write(int fd, uint16_t command, int len, void *token);
+extern int ipc_child_write(int fd, uint16_t command, size_t len, void *token);
 extern int ipc_child_state(AFPObj *obj, uint16_t state);
 
 #endif /* IPC_GETSESSION_LOGIN */
